requirements.txt
requests>=2.20
python-dotenv
openai
markdownify
mcp.server